+++ doc/sphinx/src/news.rst
@@ -23,6 +23,7 @@ New features
 * Execution
 
   * ``params_tag`` is a base class for user-defined classes that are a composition of task-parameter types (like field accessors).
+  * ``arg_tag`` is a base class for user-defined classes that specifiy custom handling as a task argument.
 
 Changes in v2.4.1
 +++++++++++++++++